Handle "reexec"-based subprocessing
In docker, finding and applying layer diffs is done by subprocesses
which are forked and exec'd with different argv[0] values, in much the
same way busybox and various other tools use argv[0] to sometimes behave
as different programs.

Here we wire in the chrootarchive hooks and the reexec module, and give
it a chance to run the alternate main functions before we actually start
acting like ourselves.

There are reexec hooks in the execdriver logic, but since we're not in
the business of running containers, we leave them out.